test: Expand tests for complex type.

+++ test/vast/Dialect/HighLevel/complex-d.c
@@ -4,7 +4,7 @@
 
 void fun() {
     double complex x = 3;
-    double y = 3;
+    float y = 3;
 // CHECK: hl.fcmp {{.*}} : !hl.complex<!hl.double>, !hl.double
     int cmp1 = x == y;
 // CHECK: hl.fcmp {{.*}} : !hl.double, !hl.complex<!hl.double>
